feat: social field common file to extend
